+++ client/src/components/Pages/NotFound/index.js
@@ -31,3 +31,4 @@ export default function NotFound() {
   );
 }
 
+// UPDATE FOR NEW STYLE
